# Improve lockmap write benchmark