Previously, it was mostly benchmarking `random.String(64)`. I pulled that out of the loop by pre-creating keys. I also made the number of keys variable, so we can have benchmarks for high and low contention.

I didn't modify the read benchmark because I think it's good to have one benchmark that has no contention.

Current results on my machine:
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor            
                 │  /tmp/base  │
                 │   sec/op    │
LockQPS/1-24       138.2n ± 7%
LockQPS/10-24      48.45n ± 5%
LockQPS/1000-24    24.97n ± 1%
LockQPS/10000-24   17.18n ± 1%
RLockQPS-24        17.33n ± 1%
geomean            34.63n

                 │ /tmp/base  │
                 │    B/op    │
LockQPS/1-24       64.00 ± 0%
LockQPS/10-24      64.00 ± 0%
LockQPS/1000-24    64.00 ± 0%
LockQPS/10000-24   64.00 ± 0%
RLockQPS-24        64.00 ± 0%
geomean            64.00

                 │ /tmp/base  │
                 │ allocs/op  │
LockQPS/1-24       3.000 ± 0%
LockQPS/10-24      3.000 ± 0%
LockQPS/1000-24    3.000 ± 0%
LockQPS/10000-24   3.000 ± 0%
RLockQPS-24        3.000 ± 0%
geomean            3.000
```