@vanja-p 2025-05-19
BTW, after this benchmark change, here are the results for Fabian's change (https://github.com/buildbuddy-io/buildbuddy/pull/9349)
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor            
                 │  /tmp/before  │             /tmp/after             │
                 │    sec/op     │   sec/op     vs base               │
LockQPS/1-24        172.4n ± 13%   138.2n ± 7%  -19.84% (p=0.002 n=6)
LockQPS/10-24      184.15n ± 11%   48.45n ± 5%  -73.69% (p=0.002 n=6)
LockQPS/1000-24    232.30n ± 13%   24.97n ± 1%  -89.25% (p=0.002 n=6)
LockQPS/10000-24   219.70n ± 15%   17.18n ± 1%  -92.18% (p=0.002 n=6)
RLockQPS-24        148.75n ± 11%   17.33n ± 1%  -88.35% (p=0.002 n=6)
geomean             189.0n         34.63n       -81.68%

                 │ /tmp/before │             /tmp/after            │
                 │    B/op     │    B/op     vs base               │
LockQPS/1-24        72.00 ± 0%   64.00 ± 0%  -11.11% (p=0.002 n=6)
LockQPS/10-24       72.00 ± 0%   64.00 ± 0%  -11.11% (p=0.002 n=6)
LockQPS/1000-24     72.00 ± 0%   64.00 ± 0%  -11.11% (p=0.002 n=6)
LockQPS/10000-24    74.00 ± 1%   64.00 ± 0%  -13.51% (p=0.002 n=6)
RLockQPS-24         72.00 ± 0%   64.00 ± 0%  -11.11% (p=0.002 n=6)
geomean             72.40        64.00       -11.60%

                 │ /tmp/before │             /tmp/after            │
                 │  allocs/op  │ allocs/op   vs base               │
LockQPS/1-24        4.000 ± 0%   3.000 ± 0%  -25.00% (p=0.002 n=6)
LockQPS/10-24       4.000 ± 0%   3.000 ± 0%  -25.00% (p=0.002 n=6)
LockQPS/1000-24     4.000 ± 0%   3.000 ± 0%  -25.00% (p=0.002 n=6)
LockQPS/10000-24    4.000 ± 0%   3.000 ± 0%  -25.00% (p=0.002 n=6)
RLockQPS-24         4.000 ± 0%   3.000 ± 0%  -25.00% (p=0.002 n=6)
geomean             4.000        3.000       -25.00%
```